+++ onnx/common/assertions.h
